requirements.txt
# Core dependencies
streamlit>=1.22.0
numpy>=1.24.3
pandas>=2.0.1
pillow>=9.5.0

# NLP and PII detection
spacy>=3.5.3
transformers>=4.29.2
pytesseract>=0.3.10

# File format handling
python-docx>=0.8.11
PyMuPDF>=1.22.3  # For PDF processing
opencv-python>=4.7.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0